feat: check for insufficient git history (#201)

+++ module/semver/src/main/java/com/xenoterracide/gradle/semver/GitMetadataExtensionAdapter.java
@@ -26,7 +26,7 @@ class GitMetadataExtensionAdapter implements GitMetadata {
   }
 
   @Override
-  public int distance() {
+  public long distance() {
     return this.delegate.getDistance().get();
   }
