+++ module/semver/src/main/java/com/xenoterracide/gradle/semver/internal/GitMetadata.java
@@ -33,7 +33,7 @@ public interface GitMetadata {
    *
    * @return the commit distance
    */
-  int distance();
+  long distance();
 
   /**
    * Gets status.
